ecxluded fxi

+++ schema.sql
@@ -34,8 +34,8 @@ CREATE TABLE birthdays (
     id SERIAL PRIMARY KEY,
     -- either discord id or custom
     user_id BIGINT NULL REFERENCES users(user_id) ON DELETE CASCADE,
-    day TINYINT NOT NULL,
-    month TINYINT NOT NULL,
+    day SMALLINT NOT NULL,
+    month SMALLINT NOT NULL,
     year SMALLINT NULL,
     timezone VARCHAR(50) NULL,
     UNIQUE(user_id)
